# pkg/server/docs/swagger.yaml
definitions:
  server.HTTPError:
    properties:
      message: {}
    type: object
  storage.StatsCounterDto:
    properties:
      count:
        type: integer
      stats:
        items:
          $ref: '#/definitions/storage.StatsDto'
        type: array
    type: object
  storage.StatsDto:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      int1:
        type: integer
      int2:
        type: integer
      limit:
        type: integer
      result:
        type: string
      str1:
        type: string
      str2:
        type: string
    type: object
info:
  contact: {}
paths:
  /fizz-buzz:
    get:
      description: 'The original fizz-buzz consists in writing all numbers from `1`
        to `100`, and replacing all multiples of `3` by `fizz`, all multiples of `5`
        by `buzz`, and all multiples of `15` by `fizzbuzz`. The output would look
        like this: 1,2,fizz,4,buzz,fizz,7,8,fizz,buzz,11,fizz,13,14,fizzbuzz,fizz,...'
      operationId: launch-fizz-buzz
      parameters:
      - description: First string related to multiples of int1
        in: query
        name: string1
        type: string
      - description: Second string related to multiples of int2
        in: query
        name: string2
        type: string
      - description: multiple related to string1
        in: query
        name: int1
        required: true
        type: integer
      - description: multiple related to string2
        in: query
        name: int2
        required: true
        type: integer
      - description: number of elements to return
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Launches the game with the required parameters
      tags:
      - fizz-buzz
  /stats:
    get:
      description: Everytime the fizz-buzz endpoint is called, the parameters related
        to the call are saved in a database, along with the time of the call, and
        the result given to the user. Calling this endpoint will return all statistics
        about the calls made so far.
      operationId: get-fizz-buzz-stats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/storage.StatsCounterDto'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Returns the parameters and result of the last calls to the fizz-buzz
        endpoint
      tags:
      - fizz-buzz
swagger: "2.0"